refactor(templates): change name
