+++ src/components/templates/index/index.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react'
 import style from '../index.module.scss'
 
-import Section from '../_utils/Section'
+import Section from '../_layout/Section'
 
 import HeaderProfile from '../../organisms/HeaderProfile'
